Created a branch to try and restructure the project
